🎨 refactor: refactor the route auth as a middleware
