+++ src/app/api/chat/auth/utils.ts
